+++ src/main/java/io/github/jamalam360/rightclickharvest/mixin/NetherWartBlockMixin.java
@@ -44,10 +44,20 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public abstract class NetherWartBlockMixin extends AbstractBlockMixin {
     @Override
     public void rightClickHarvest(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit, CallbackInfoReturnable<ActionResult> info) {
+        if (Config.useHunger) {
+            if (player.getHungerManager().getFoodLevel() <= 0) {
+                return;
+            }
+        }
+
         if (state.get(NetherWartBlock.AGE) >= NetherWartBlock.MAX_AGE) {
             if (!world.isClient) {
                 world.setBlockState(pos, state.with(NetherWartBlock.AGE, 0));
                 Block.dropStacks(state, world, pos, null, player, player.getStackInHand(hand));
+
+                if (Config.useHunger && player.world.random.nextBoolean()) {
+                    player.addExhaustion(2f);
+                }
             } else {
                 player.playSound(SoundEvents.ITEM_NETHER_WART_PLANT, 1.0f, 1.0f);
             }